refactor: extract insight net worth chart
