+++ frontend/src/insights/components/NetWorthCard.tsx
@@ -1,25 +1,7 @@
-import {
-  useMemo,
-  useRef,
-  type MouseEvent as ReactMouseEvent,
-} from 'react'
-import { AnimatePresence, motion } from 'motion/react'
+import { useMemo } from 'react'
 import { ArrowLeftRight, Minus, TrendingDown, TrendingUp, Wallet } from 'lucide-react'
-import {
-  Area,
-  Bar,
-  ComposedChart,
-  Line,
-  ReferenceDot,
-  ReferenceLine,
-  ResponsiveContainer,
-  XAxis,
-  YAxis,
-} from 'recharts'
 import type { FxStatus } from '@/api/shared/fx'
-import { ChartTooltipRow, ChartTooltipTitle } from '@/components/charts/ChartTooltipContent'
 import IconTooltip from '@/components/IconTooltip'
-import { DASHBOARD_X_AXIS_TICK_FONT_SIZE } from '@/dashboard/constants/chart'
 import { getInsightsNetWorthFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
 import { formatCurrency } from '@/utils/formatCurrency'
 import { FxStatusBadge } from './FxStatusBadge'
@@ -27,44 +9,20 @@ import {
   InsightLoadingContent,
   InsightLoadingOverlay,
 } from './InsightLoadingTransition'
-import {
-  DeferredChartTooltipOverlay,
-  type ChartTooltipPointer,
-  type DeferredChartTooltipOverlayHandle,
-} from '@/components/charts/DeferredChartTooltipOverlay'
 import { InsightActionButton } from './InsightActionButton'
+import { NetWorthChart } from './net-worth/NetWorthChart'
 import { SectionHeader } from './SectionHeader'
 import { useInsightLoadingSnapshot } from './useInsightLoadingSnapshot'
 import {
-  NET_WORTH_AXIS_TICK_COUNT,
-  formatNetWorthAxisDate,
-  formatNetWorthAxisMoney,
   formatSignedNetWorthCurrency,
-  getChangeKey,
-  getChartKey,
   getNetWorthChartData,
   getNetWorthChartItems,
-  getNetWorthDateAxisTicks,
-  getNetWorthLegendItems,
-  getValueKey,
-  netWorthChangeColor,
-  netWorthChartLeftMargin,
-  type NetWorthChartItem,
-  type NetWorthDeltaPoint,
   type NetWorthGroup,
   type NetWorthPoint,
   type NetWorthViewMode,
 } from '../utils/netWorthChart'
 
 export type { NetWorthViewMode } from '../utils/netWorthChart'
-
-type AxisTickProps = {
-  x?: number | string
-  y?: number | string
-  payload?: {
-    value: number | string
-  }
-}
 
 type NetWorthCardProps = {
   mode: NetWorthViewMode
@@ -88,146 +46,18 @@ type NetWorthSnapshot = {
   emptyLabel: string
 }
 
-type NetWorthTooltipState = {
-  activeLabel?: string | number
-  activeCoordinate?: {
-    x?: number
-  }
-}
-
-const netWorthLegendContainerVariants = {
-  initial: { transition: { staggerChildren: 0.035, staggerDirection: 1 } },
-  enter: { transition: { staggerChildren: 0.045, staggerDirection: 1, delayChildren: 0.03 } },
-  exit: { transition: { staggerChildren: 0.035, staggerDirection: 1 } },
-} as const
-
-const netWorthLegendItemVariants = {
-  initial: { opacity: 0, x: -10 },
-  enter: { opacity: 1, x: 0 },
-  exit: { opacity: 0, x: 10 },
-} as const
-
-const netWorthChartMargin = { top: 18, right: 0, bottom: 0, left: netWorthChartLeftMargin } as const
-const netWorthXAxisPadding = { left: 28, right: 28 } as const
-const netWorthLegendItemTransition = { duration: 0.22, ease: [0.16, 1, 0.3, 1] } as const
-
-function getNetWorthTooltipKey(point: NetWorthDeltaPoint) {
-  return point.dateMs
-}
-
+/**
+ * Describes the calculation currently shown by the net-worth card mode
+ */
 function getNetWorthCalculation(mode: NetWorthViewMode) {
   return mode === 'overview'
     ? 'Net worth is assets minus debt. Change compares with net worth from the day before this range'
     : 'Balances are grouped by account type at each chart date'
 }
 
-function getNetWorthFirstBarLabelY(point: NetWorthDeltaPoint, items: NetWorthChartItem[]) {
-  const positiveStack = items.reduce((sum, _item, index) => {
-    const value = Number(point[getChartKey(index)] ?? 0)
-    return value > 0 ? sum + value : sum
-  }, 0)
-
-  return positiveStack > 0 ? positiveStack : 0
-}
-
-function getNetWorthTooltipPointer(
-  state: NetWorthTooltipState,
-  event: ReactMouseEvent<SVGGraphicsElement>,
-): ChartTooltipPointer {
-  return {
-    clientX: event.clientX,
-    clientY: event.clientY,
-    chartX: typeof state.activeCoordinate?.x === 'number' ? state.activeCoordinate.x : undefined,
-  }
-}
-
-function NetWorthXAxisTick({
-  x = 0,
-  y = 0,
-  payload,
-  dateLabelsByMs,
-}: AxisTickProps & {
-  dateLabelsByMs: Map<number, string>
-}) {
-  const value = Number(payload?.value)
-  const tickX = Number(x)
-  const tickY = Number(y)
-
-  return (
-    <text
-      x={tickX}
-      y={tickY}
-      dy={12}
-      textAnchor="middle"
-      fill="var(--app-text-subtle)"
-      fontSize={DASHBOARD_X_AXIS_TICK_FONT_SIZE}
-    >
-      {dateLabelsByMs.get(value) ?? formatNetWorthAxisDate(value)}
-    </text>
-  )
-}
-
-function NetWorthChartTooltipContent({
-  point,
-  items,
-  displayCurrency,
-  mode,
-}: {
-  point: NetWorthDeltaPoint
-  items: NetWorthChartItem[]
-  displayCurrency: string
-  mode: NetWorthViewMode
-}) {
-  const detailItems = items.map((item, index) => ({ item, index }))
-  const displayedNetWorth = mode === 'composition'
-    ? detailItems.reduce((sum, { index }) => sum + Number(point[getValueKey(index)] ?? 0), 0)
-    : point.total
-
-  return (
-    <>
-      <ChartTooltipTitle>{point.tooltipLabel}</ChartTooltipTitle>
-      <ChartTooltipRow
-        label="Net Worth"
-        value={formatCurrency(displayedNetWorth, displayCurrency)}
-        financialValue
-      />
-      {mode === 'overview' && (
-        <ChartTooltipRow
-          label="Change"
-          value={formatSignedNetWorthCurrency(point.totalChange, displayCurrency)}
-          financialValue
-        />
-      )}
-      <div className="mt-2 space-y-1 border-t border-[var(--app-border)] pt-2">
-        {detailItems.map(({ item, index }) => {
-          const value = Number(point[getValueKey(index)] ?? 0)
-          const change = Number(point[getChangeKey(index)] ?? 0)
-          const displayValue = item.kind === 'debt' ? Math.abs(value) : value
-          return (
-            <ChartTooltipRow
-              key={item.id}
-              className="mt-0"
-              label={item.name}
-              value={(
-                <>
-                  {formatCurrency(displayValue, displayCurrency)}
-                  {mode === 'overview' && (
-                    <>
-                      {' '}
-                      ({formatSignedNetWorthCurrency(change, displayCurrency)})
-                    </>
-                  )}
-                </>
-              )}
-              financialValue
-            />
-          )
-        })}
-      </div>
-    </>
-  )
-}
-
+/**
+ * Renders net-worth summary metrics and the net-worth history chart
+ */
 export function NetWorthCard({
   mode,
   onModeToggle,
@@ -239,8 +69,6 @@ export function NetWorthCard({
   loading = false,
   transitionKey,
 }: NetWorthCardProps) {
-  const netWorthChartRef = useRef<HTMLDivElement>(null)
-  const netWorthTooltipRef = useRef<DeferredChartTooltipOverlayHandle<NetWorthDeltaPoint>>(null)
   const incomingSnapshot = useMemo<NetWorthSnapshot>(() => ({
     mode,
     groups,
@@ -269,34 +97,6 @@ export function NetWorthCard({
     () => getNetWorthChartData(displaySnapshot.series, chartItems, displaySnapshot.mode, displaySnapshot.baseline),
     [chartItems, displaySnapshot.baseline, displaySnapshot.mode, displaySnapshot.series],
   )
-  const hasChartData = displaySnapshot.groups.length > 0 && deltaSeries.length > 0
-  const dateAxisStartMs = deltaSeries[0]?.dateMs ?? 0
-  const dateAxisEndMs = deltaSeries.at(-1)?.dateMs ?? dateAxisStartMs
-  const dateAxisTicks = useMemo(
-    () => getNetWorthDateAxisTicks(deltaSeries, NET_WORTH_AXIS_TICK_COUNT),
-    [deltaSeries],
-  )
-  const dateLabelsByMs = useMemo(
-    () => new Map(deltaSeries.map((point) => [point.dateMs, point.dateLabel])),
-    [deltaSeries],
-  )
-  const netWorthPointsByMs = useMemo(
-    () => new Map(deltaSeries.map((point) => [point.dateMs, point])),
-    [deltaSeries],
-  )
-  const startNetWorthAxisLabel = formatNetWorthAxisMoney(
-    deltaSeries[0]?.startTotal ?? 0,
-    displaySnapshot.displayCurrency,
-  )
-  const startNetWorthLabelPoint = displaySnapshot.mode === 'overview' ? deltaSeries[0] : undefined
-  const startNetWorthLabelY = startNetWorthLabelPoint
-    ? getNetWorthFirstBarLabelY(startNetWorthLabelPoint, chartItems)
-    : 0
-  const legendItems = useMemo(
-    () => getNetWorthLegendItems(displaySnapshot.mode, chartItems),
-    [chartItems, displaySnapshot.mode],
-  )
-  const legendAnimationKey = `${displaySnapshot.mode}-${legendItems.map((item) => item.id).join('|')}`
   const latestChange = deltaSeries.at(-1)?.totalChange ?? 0
   const netWorthTrendColor = latestChange > 0
     ? 'var(--app-chart-positive)'
@@ -304,21 +104,6 @@ export function NetWorthCard({
       ? 'var(--app-chart-negative)'
       : 'var(--app-text-muted)'
   const NetWorthTrendIcon = latestChange > 0 ? TrendingUp : latestChange < 0 ? TrendingDown : Minus
-  const showNetWorthTooltip = (
-    state: NetWorthTooltipState,
-    event: ReactMouseEvent<SVGGraphicsElement>,
-  ) => {
-    const activeDateMs = Number(state.activeLabel)
-    const point = Number.isFinite(activeDateMs) ? netWorthPointsByMs.get(activeDateMs) : undefined
-    const pointer = getNetWorthTooltipPointer(state, event)
-    if (!point) {
-      netWorthTooltipRef.current?.show(null, pointer)
-      return
-    }
-
-    netWorthTooltipRef.current?.show(point, pointer)
-  }
-  const hideNetWorthTooltip = () => netWorthTooltipRef.current?.hide()
 
   return (
     <section className="app-card">
@@ -358,186 +143,41 @@ export function NetWorthCard({
       <div className="relative overflow-visible" data-tooltip-bounds>
         <InsightLoadingContent concealed={contentConcealed} shouldReduceMotion={shouldReduceMotion}>
           <div className="flex h-[360px] flex-col">
-        <div className="mb-3 flex flex-wrap items-end justify-between gap-3">
-          <div>
-            <p className="app-label app-label-compact inline-flex items-center gap-2">
-              Ending Net Worth
-              <IconTooltip
-                label="Ending Net Worth calculation"
-                placement="top"
-                widthClassName="w-72"
-                size={14}
-                strokeWidth={2.25}
-              >
-                Ending net worth value as of the last date in the chosen time period
-              </IconTooltip>
-            </p>
-            <div className="mt-1 flex flex-wrap items-end gap-x-2 gap-y-1">
-              <p className="font-financial text-3xl leading-none tracking-tight">
-                {formatCurrency(latest?.total ?? 0, displaySnapshot.displayCurrency)}
-              </p>
-              <div className="flex items-center gap-1.5 text-sm font-medium leading-none" style={{ color: netWorthTrendColor }}>
-                <NetWorthTrendIcon size={14} aria-hidden />
-                <span className="font-financial">{formatSignedNetWorthCurrency(latestChange, displaySnapshot.displayCurrency)}</span>
-                <span style={{ color: 'var(--app-text-subtle)' }}>since start</span>
+            <div className="mb-3 flex flex-wrap items-end justify-between gap-3">
+              <div>
+                <p className="app-label app-label-compact inline-flex items-center gap-2">
+                  Ending Net Worth
+                  <IconTooltip
+                    label="Ending Net Worth calculation"
+                    placement="top"
+                    widthClassName="w-72"
+                    size={14}
+                    strokeWidth={2.25}
+                  >
+                    Ending net worth value as of the last date in the chosen time period
+                  </IconTooltip>
+                </p>
+                <div className="mt-1 flex flex-wrap items-end gap-x-2 gap-y-1">
+                  <p className="font-financial text-3xl leading-none tracking-tight">
+                    {formatCurrency(latest?.total ?? 0, displaySnapshot.displayCurrency)}
+                  </p>
+                  <div className="flex items-center gap-1.5 text-sm font-medium leading-none" style={{ color: netWorthTrendColor }}>
+                    <NetWorthTrendIcon size={14} aria-hidden />
+                    <span className="font-financial">{formatSignedNetWorthCurrency(latestChange, displaySnapshot.displayCurrency)}</span>
+                    <span style={{ color: 'var(--app-text-subtle)' }}>since start</span>
+                  </div>
+                </div>
               </div>
             </div>
-          </div>
-        </div>
-        <div
-          ref={netWorthChartRef}
-          className="relative min-h-0 flex-1"
-          onMouseLeave={hideNetWorthTooltip}
-        >
-          {!hasChartData ? (
-            <div
-              className="flex h-full w-full items-center justify-center rounded-lg text-sm"
-              style={{ color: 'var(--app-text-subtle)' }}
-            >
-              {displaySnapshot.emptyLabel}
-            </div>
-          ) : (
-            <ResponsiveContainer width="100%" height="100%">
-              <ComposedChart
-                data={deltaSeries}
-                margin={netWorthChartMargin}
-                onMouseMove={(state, event) => showNetWorthTooltip(state, event)}
-                onMouseLeave={hideNetWorthTooltip}
-              >
-                <XAxis
-                  dataKey="dateMs"
-                  type="number"
-                  scale="time"
-                  domain={[dateAxisStartMs, dateAxisEndMs]}
-                  padding={netWorthXAxisPadding}
-                  ticks={dateAxisTicks}
-                  axisLine={false}
-                  tickLine={false}
-                  interval={0}
-                  tick={(props) => (
-                    <NetWorthXAxisTick
-                      {...props}
-                      dateLabelsByMs={dateLabelsByMs}
-                    />
-                  )}
-                  tickMargin={4}
-                />
-                <YAxis
-                  hide
-                  domain={[(dataMin: number) => Math.min(dataMin, 0), (dataMax: number) => Math.max(dataMax, 0)]}
-                />
-                <ReferenceLine
-                  y={0}
-                  stroke="var(--app-border-strong)"
-                  strokeWidth={1}
-                />
-                {startNetWorthLabelPoint && (
-                  <ReferenceDot
-                    x={startNetWorthLabelPoint.dateMs}
-                    y={startNetWorthLabelY}
-                    r={0}
-                    fill="transparent"
-                    stroke="transparent"
-                    ifOverflow="visible"
-                    label={{
-                      value: startNetWorthAxisLabel,
-                      position: 'top',
-                      offset: 4,
-                      fill: 'var(--app-text-subtle)',
-                      fontSize: 11,
-                      fontWeight: 600,
-                    }}
-                  />
-                )}
-                {displaySnapshot.mode === 'composition' ? (
-                  chartItems.map((item, index) => (
-                    <Area
-                      key={item.id}
-                      type="monotone"
-                      dataKey={getChartKey(index)}
-                      stackId={item.kind}
-                      stroke={item.color}
-                      strokeWidth={1.5}
-                      fill={item.color}
-                      fillOpacity={0.65}
-                      activeDot={false}
-                    />
-                  ))
-                ) : (
-                  <>
-                    {chartItems.map((item, index) => (
-                      <Bar
-                        key={item.id}
-                        dataKey={getChartKey(index)}
-                        stackId="net-worth-contribution"
-                        fill={item.color}
-                        maxBarSize={34}
-                        radius={4}
-                      />
-                    ))}
-                    <Line
-                      type="monotone"
-                      dataKey="totalChange"
-                      stroke={netWorthChangeColor}
-                      strokeWidth={2.5}
-                      dot={false}
-                      activeDot={{ r: 4, strokeWidth: 0 }}
-                    />
-                  </>
-                )}
-              </ComposedChart>
-            </ResponsiveContainer>
-          )}
-          {hasChartData && (
-            <DeferredChartTooltipOverlay
-              ref={netWorthTooltipRef}
-              chartRef={netWorthChartRef}
-              className="min-w-64"
-              getKey={getNetWorthTooltipKey}
-              renderContent={(point) => (
-                <NetWorthChartTooltipContent
-                  point={point}
-                  items={chartItems}
-                  displayCurrency={displaySnapshot.displayCurrency}
-                  mode={displaySnapshot.mode}
-                />
-              )}
+            <NetWorthChart
+              mode={displaySnapshot.mode}
+              groups={displaySnapshot.groups}
+              chartItems={chartItems}
+              deltaSeries={deltaSeries}
+              displayCurrency={displaySnapshot.displayCurrency}
+              emptyLabel={displaySnapshot.emptyLabel}
+              shouldReduceMotion={shouldReduceMotion}
             />
-          )}
-        </div>
-        <div className="mt-3 overflow-hidden">
-          <AnimatePresence initial={false} mode="wait">
-            {hasChartData && (
-              <motion.div
-                key={legendAnimationKey}
-                className="flex flex-wrap justify-center gap-x-4 gap-y-1"
-                variants={shouldReduceMotion ? undefined : netWorthLegendContainerVariants}
-                initial={shouldReduceMotion ? false : 'initial'}
-                animate={shouldReduceMotion ? { opacity: 1 } : 'enter'}
-                exit={shouldReduceMotion ? undefined : 'exit'}
-              >
-                {legendItems.map((item) => (
-                  <motion.div
-                    key={item.id}
-                    className="flex items-center gap-1.5 text-xs"
-                    style={{ color: 'var(--app-text-muted)' }}
-                    variants={shouldReduceMotion ? undefined : netWorthLegendItemVariants}
-                    transition={shouldReduceMotion ? { duration: 0 } : netWorthLegendItemTransition}
-                  >
-                    <span
-                      className="h-2.5 w-2.5 rounded-full"
-                      style={{
-                        background: item.color,
-                        opacity: displaySnapshot.mode === 'composition' ? 0.65 : 1,
-                      }}
-                    />
-                    <span>{item.name}</span>
-                  </motion.div>
-                ))}
-              </motion.div>
-            )}
-          </AnimatePresence>
-        </div>
           </div>
         </InsightLoadingContent>
